reverted frontent integration - code extracted to a new test
