+++ apps/app/src/test/resources/localnet-reassign-topology.conf
@@ -1,0 +1,38 @@
+include required("include/canton-basic.conf")
+canton {
+  validator-app-clients {
+    userValidatorClient {
+      admin-api {
+        url = "http://wallet.localhost:2000"
+      }
+    }
+    providerValidatorClient {
+      admin-api {
+        url = "http://wallet.localhost:3000"
+      }
+    }
+  }
+  remote-participants {
+    app-provider {
+      ledger-api {
+        address = "grpc-ledger-api.localhost"
+        port = 3000
+      }
+      admin-api {
+        address = "localhost"
+        port = 3902
+      }
+    }
+    app-user {
+      ledger-api {
+        address = "grpc-ledger-api.localhost"
+        port = 2000
+      }
+      admin-api {
+        address = "localhost"
+        port = 2902
+      }
+    }
+  }
+}
+
