+++ apps/app/src/test/scala/org/lfdecentralizedtrust/splice/integration/tests/LocalNetFrontendIntegrationTest.scala
@@ -11,24 +11,11 @@ import org.lfdecentralizedtrust.splice.console.ParticipantClientReference
 import org.lfdecentralizedtrust.splice.integration.EnvironmentDefinition
 import org.lfdecentralizedtrust.splice.util.{FrontendLoginUtil, WalletFrontendTestUtil}
 import com.digitalasset.canton.http.json.v2.JsStateServiceCodecs.*
-import com.digitalasset.canton.protocol.LfContractId
-import com.digitalasset.canton.topology.SynchronizerId
 import com.digitalasset.canton.version.ProtocolVersion
 import com.daml.ledger.api.v2.state_service.GetConnectedSynchronizersResponse
-import com.daml.ledger.api.v2.transaction_filter.CumulativeFilter.IdentifierFilter
-import com.daml.ledger.api.v2.transaction_filter.{
-  CumulativeFilter,
-  EventFormat,
-  Filters,
-  WildcardFilter,
-}
 import monocle.Monocle.toAppliedFocusOps
-import org.lfdecentralizedtrust.splice.codegen.java.splice.api.token.test.dummyholding.DummyHolding
-import org.lfdecentralizedtrust.splice.util.JavaDecodeUtil
 
-import java.nio.file.Paths
 import scala.concurrent.duration.*
-import scala.jdk.CollectionConverters.*
 import scala.sys.process.*
 
 class LocalNetFrontendIntegrationTest
@@ -41,8 +28,7 @@ class LocalNetFrontendIntegrationTest
       .updateTestingConfig(
         _.focus(_.participantsWithoutLapiVerification).replace(
           Set(
-            "app-provider",
-            "app-user",
+            "app-provider"
           )
         )
       )
@@ -54,10 +40,6 @@ class LocalNetFrontendIntegrationTest
   override lazy val resetRequiredTopologyState = false
 
   val partyHint = "localnet-localparty-1"
-
-  // The user all localnet nodes use for their ledger API access, see
-  // cluster/compose/localnet/env/*-auth-on.env
-  private val ledgerApiUserId = "ledger-api-user"
 
   private def withLocalNet(
       additionalArgs: Seq[String]
@@ -117,9 +99,7 @@ class LocalNetFrontendIntegrationTest
       withFrontEnd("frontend") { implicit webDriver =>
         actAndCheck(
           "Open the Scan UI",
-          // The scheme is required: without it Firefox parses `scan.localhost:` as a
-          // (non-special) URL scheme and refuses to navigate.
-          go to "http://scan.localhost:4000",
+          go to "scan.localhost:4000",
         )(
           "Open rounds should be listed",
           _ => findAll(className("open-mining-round-row")) should have length 2,
@@ -148,14 +128,7 @@ class LocalNetFrontendIntegrationTest
       }
     }
 
-  private val token = AuthUtil.testToken(AuthUtil.testAudience, ledgerApiUserId, "unsafe")
-
-  private val dummyHoldingDarPath = Paths
-    .get(
-      "token-standard/examples/splice-token-test-dummy-holding/.daml/dist/splice-token-test-dummy-holding-current.dar"
-    )
-    .toAbsolutePath
-    .toString
+  private val token = AuthUtil.testToken(AuthUtil.testAudience, "ledger-api-user", "unsafe")
 
   private def testTokenStandardApi(implicit env: FixtureParam): Unit =
     clue("Test token standard APIs") {
@@ -202,120 +175,6 @@ class LocalNetFrontendIntegrationTest
       }
     }
 
-  private def participantClient(name: String)(implicit env: FixtureParam) = {
-    val remoteParticipant =
-      env.participants.remote
-        .find(_.name == name)
-        .getOrElse(fail(s"$name participant not found"))
-    new ParticipantClientReference(
-      env,
-      remoteParticipant.name,
-      remoteParticipant.config.copy(token = Some(token)),
-    )
-  }
-
-  private def synchronizerId(
-      participant: ParticipantClientReference,
-      alias: String,
-  ): SynchronizerId =
-    participant.synchronizers
-      .list_connected()
-      .find(_.synchronizerAlias.unwrap == alias)
-      .getOrElse(fail(s"${participant.name} is not connected to $alias"))
-      .synchronizerId
-
-  private def testReassignment(participantName: String, validatorClientName: String)(implicit
-      env: FixtureParam
-  ): Unit =
-    clue(s"Reassign a contract between global and app-synchronizer on $participantName") {
-      val participant = participantClient(participantName)
-      val party = vc(validatorClientName).copy(token = Some(token)).getValidatorPartyId()
-      val globalSynchronizerId = synchronizerId(participant, "global")
-      val appSynchronizerId = synchronizerId(participant, "app-synchronizer")
-
-      participant.upload_dar_unless_exists(dummyHoldingDarPath)
-
-      val createdContract = clue("Create a DummyHolding on the global synchronizer") {
-        val tx = participant.ledger_api_extensions.commands.submitJava(
-          actAs = Seq(party),
-          commands = new DummyHolding(
-            party.toProtoPrimitive,
-            party.toProtoPrimitive,
-            BigDecimal(42).bigDecimal,
-          ).create().commands().asScala.toSeq,
-          synchronizerId = Some(globalSynchronizerId),
-          userId = ledgerApiUserId,
-        )
-        JavaDecodeUtil.decodeAllCreated(DummyHolding.COMPANION)(tx).loneElement
-      }
-      val contractId = createdContract.id.contractId
-      val lfContractId = LfContractId.assertFromString(contractId)
-
-      def synchronizerOfContract() =
-        participant.ledger_api_extensions.acs
-          .lookup_contract_domain(party, Set(contractId))
-          .get(contractId)
-
-      synchronizerOfContract() should be(Some(globalSynchronizerId))
-
-      val eventFormat = Some(
-        EventFormat(
-          filtersByParty = Map(
-            party.toProtoPrimitive -> Filters(
-              Seq(
-                CumulativeFilter(
-                  IdentifierFilter.WildcardFilter(
-                    WildcardFilter(includeCreatedEventBlob = false)
-                  )
-                )
-              )
-            )
-          ),
-          filtersForAnyParty = None,
-          verbose = true,
-        )
-      )
-
-      def reassign(source: SynchronizerId, target: SynchronizerId): Unit = {
-        val unassigned = participant.ledger_api.commands
-          .submit_unassign_with_format(
-            submitter = party,
-            contractIds = Seq(lfContractId),
-            source = source,
-            target = target,
-            userId = ledgerApiUserId,
-            eventFormat = eventFormat,
-            timeout = None,
-          )
-          .unassignedWrapper
-        val _ = participant.ledger_api.commands.submit_assign_with_format(
-          submitter = party,
-          reassignmentId = unassigned.reassignmentId,
-          source = source,
-          target = target,
-          userId = ledgerApiUserId,
-          eventFormat = eventFormat,
-          timeout = None,
-        )
-      }
-
-      actAndCheck(
-        "Reassign the contract to the app-synchronizer",
-        reassign(globalSynchronizerId, appSynchronizerId),
-      )(
-        "The contract is now assigned to the app-synchronizer",
-        _ => synchronizerOfContract() should be(Some(appSynchronizerId)),
-      )
-
-      actAndCheck(
-        "Reassign the contract back to the global synchronizer",
-        reassign(appSynchronizerId, globalSynchronizerId),
-      )(
-        "The contract is assigned to the global synchronizer again",
-        _ => synchronizerOfContract() should be(Some(globalSynchronizerId)),
-      )
-    }
-
   "docker-compose based localnet works for single synchronizer" in { implicit env =>
     withLocalNet(Seq.empty) { implicit env =>
       testValidators
@@ -328,15 +187,22 @@ class LocalNetFrontendIntegrationTest
   "docker-compose based localnet works for multiple synchronizers" in { implicit env =>
     withLocalNet(Seq("-M")) { implicit env =>
       testMultiSynchronizerSupport(isMultiSync = true)
-      testReassignment("app-provider", "providerValidatorClient")
-      testReassignment("app-user", "userValidatorClient")
     }
   }
 
   "localnet supports configurable protocol versions" in { implicit env =>
     withLocalNet(Seq("-u", "-p", "35")) { implicit env =>
-      val appProviderParticipant = participantClient("app-provider")
-      appProviderParticipant.synchronizers
+      val appProviderParticipant =
+        env.participants.remote
+          .find(_.name == "app-provider")
+          .getOrElse(fail("app-provider participant not found"))
+
+      val participantClient = new ParticipantClientReference(
+        env,
+        appProviderParticipant.name,
+        appProviderParticipant.config.copy(token = Some(token)),
+      )
+      participantClient.synchronizers
         .list_connected()
         .loneElement
         .physicalSynchronizerId
